+++ RELEASES.md
@@ -1,5 +1,15 @@
 # Confidential Match Source Code Releases
 
+## May 29, 2025
+
+* Version: v1.162.0
+* Lookup Service
+  * Image Digest: sha256:fb58ef3830b9c89019956af438942f17fd6178f1b808b0d267aae15d861a66f3
+  * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/lookup_server_gcp_signature@sha256:64ab20b4869abebc42d5b8d3c27b290134575416691621112fd0702a0a87c378)
+* Match Request Processor
+  * Image Digest: sha256:451bab77f193abbee18cc027051576f86ea6b66f4350027af2a21b9028b446df
+  * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/mrp_app_gcp_signature@sha256:8cc9eda6dc8cda46420b1fa43d9df1b153192e1f79e4e105f898300af9695b6f)
+
 ## May 16, 2025
 
 * Version: v1.150.0